cambios switches agregado member soy un pro
