+++ registro.py
@@ -70,7 +70,7 @@ import spot
 import DataBaseLocal as DataBase
 import re
 from PIL import Image, ImageDraw
-
+import GuardadoEnDatabaseRegistro as GDBR
 Userspotify = spot.userSpot
 
 
@@ -105,7 +105,7 @@ class Registro(customtkinter.CTk):
         self.background_label.place(relx=0.5, rely=0.5, anchor=customtkinter.CENTER)
         # configure grid layout (4x4)
 
-        self.tabview = customtkinter.CTkTabview(self.background_label, width=400, height=600, fg_color="#000000",
+        self.tabview = customtkinter.CTkTabview(self.background_label, width=400, height=600, fg_color="#AFAFAF",
                                                 segmented_button_selected_color=green,
                                                 segmented_button_selected_hover_color=pink,
                                                 bg_color="transparent",corner_radius=50)  # Set bg_color to transparent
@@ -123,9 +123,9 @@ class Registro(customtkinter.CTk):
         self.tabview.tab(dic.Data[dic.language]).configure(bg_color="transparent", fg_color="transparent")
         self.tabview.tab(dic.Music[dic.language]).configure(bg_color="transparent", fg_color="transparent")
         self.tabview.tab(dic.Game[dic.language]).configure(bg_color="transparent", fg_color="transparent")
+        self.tabview.tab(dic.Members[dic.language]).configure( bg_color="transparent", fg_color="transparent")
         self.tabview.tab(dic.Palettes[dic.language]).configure(bg_color="transparent", fg_color="transparent")
         self.tabview.tab(dic.Texture[dic.language]).configure(bg_color="transparent", fg_color="transparent")
-        self.tabview.tab(dic.Members[dic.language]).configure(bg_color="transparent", fg_color="transparent")
 
         self.logo_label = customtkinter.CTkLabel(self.tabview.tab(dic.Data[dic.language]),
                                                  text=dic.Registration[dic.language],
@@ -163,12 +163,12 @@ class Registro(customtkinter.CTk):
         self.edad_button = customtkinter.CTkButton(self.tabview.tab(dic.Data[dic.language]), text="Confirmar fecha",
                                                    fg_color=green_light,
                                                    hover_color=green,
-                                                   command=lambda: [self.DateSelect(), self.toggle_calendar()])
+                                                   command=lambda: [GDBR.DateSelect(), GDBR.toggle_calendar()])
         self.edad_button.place_forget()
         self.calendario_button = customtkinter.CTkButton(self.tabview.tab(dic.Data[dic.language]),
                                                          text="Whats your birthday?",
                                                          fg_color=green_light, hover_color=green,
-                                                         command=self.toggle_calendar)
+                                                         command=GDBR.toggle_calendar)
         self.calendario_button.place(relx=0.5, rely=0.55, anchor=customtkinter.CENTER)
         # -------------------------------------------------------------------------------
 
@@ -246,7 +246,7 @@ class Registro(customtkinter.CTk):
             height=10,
             width=10,
             bg_color="transparent",  # Fondo transparente
-            command=self.abrir_archivo
+            command=GDBR.abrir_archivo
         )
         self.subir_Foto.place(relx=0.46, rely=0.65, anchor=customtkinter.CENTER)
 
@@ -261,7 +261,7 @@ class Registro(customtkinter.CTk):
             text="",
             bg_color="transparent",  # Fondo transparente
             fg_color="transparent", # Fondo transparente
-            command=self.registro_facial
+            command=GDBR.registro_facial
         )
         self.camera_button.place(relx=0.54, rely=0.65, anchor=customtkinter.CENTER)
 
@@ -270,66 +270,66 @@ class Registro(customtkinter.CTk):
         self.logo_label = customtkinter.CTkLabel(self.tabview.tab(dic.Music[dic.language]),
                                                  text=dic.Registration[dic.language],
                                                  font=customtkinter.CTkFont(size=20, weight="bold"))
-        self.logo_label.place(relx=0.5, rely=0.25, anchor=customtkinter.CENTER)
+        self.logo_label.place(relx=0.5, rely=0.1, anchor=customtkinter.CENTER)
 
         self.username = customtkinter.CTkLabel(self.tabview.tab(dic.Music[dic.language]),
                                                text=dic.FavoriteSongs[dic.language], anchor="w")
-        self.username.place(relx=0.5, rely=0.4, anchor=customtkinter.CENTER)
+        self.username.place(relx=0.5, rely=0.2, anchor=customtkinter.CENTER)
         # ---------------------------------------------------------------------------------------------
         self.SaveuserSpot = customtkinter.CTkButton(self.tabview.tab(dic.Music[dic.language]),
                                                     text="Save Spotify User",
                                                     fg_color=green_light, hover_color=green,
                                                     command=self.UserSpotSelect)
-        self.SaveuserSpot.place(relx=0.75, rely=0.3, anchor=customtkinter.CENTER)
+        self.SaveuserSpot.place(relx=0.8, rely=0.2, anchor=customtkinter.CENTER)
 
 
 
 
         self.userSpot = customtkinter.CTkEntry(self.tabview.tab(dic.Music[dic.language]), placeholder_text="Spotify User")
-        self.userSpot.place(relx=0.5, rely=0.3, anchor=customtkinter.CENTER)
+        self.userSpot.place(relx=0.5, rely=0.2, anchor=customtkinter.CENTER)
 
 
         self.cancion1 = customtkinter.CTkEntry(self.tabview.tab(dic.Music[dic.language]), placeholder_text="Song 1")
-        self.cancion1.place(relx=0.5, rely=0.35, anchor=customtkinter.CENTER)
+        self.cancion1.place(relx=0.5, rely=0.27, anchor=customtkinter.CENTER)
         self.song_button_1 = customtkinter.CTkButton(self.tabview.tab(dic.Music[dic.language]),
                                                      text="Search",
                                                      fg_color=green_light, hover_color=green,
                                                      command=self.SongSelect1)
-        self.song_button_1.place(relx=0.75, rely=0.35, anchor=customtkinter.CENTER)
+        self.song_button_1.place(relx=0.8, rely=0.27, anchor=customtkinter.CENTER)
 
         self.cancion2 = customtkinter.CTkEntry(self.tabview.tab(dic.Music[dic.language]), placeholder_text="Song 2")
-        self.cancion2.place(relx=0.5, rely=0.4, anchor=customtkinter.CENTER)
+        self.cancion2.place(relx=0.5, rely=0.34, anchor=customtkinter.CENTER)
         self.song_button_2 = customtkinter.CTkButton(self.tabview.tab(dic.Music[dic.language]),
                                                      text="Search",
                                                      fg_color=green_light, hover_color=green,
                                                      command=self.SongSelect2)
-        self.song_button_2.place(relx=0.75, rely=0.4, anchor=customtkinter.CENTER)
+        self.song_button_2.place(relx=0.8, rely=0.34, anchor=customtkinter.CENTER)
 
         self.cancion3 = customtkinter.CTkEntry(self.tabview.tab(dic.Music[dic.language]), placeholder_text="Song 3")
-        self.cancion3.place(relx=0.5, rely=0.45, anchor=customtkinter.CENTER)
+        self.cancion3.place(relx=0.5, rely=0.41, anchor=customtkinter.CENTER)
         self.song_button_3 = customtkinter.CTkButton(self.tabview.tab(dic.Music[dic.language]),
                                                      text="Search",
                                                      fg_color=green_light, hover_color=green,
                                                      command=self.SongSelect3)
-        self.song_button_3.place(relx=0.75, rely=0.45, anchor=customtkinter.CENTER)
+        self.song_button_3.place(relx=0.8, rely=0.41, anchor=customtkinter.CENTER)
         # ---------------------------------------------------------------------------------------------
         self.register_button_data = customtkinter.CTkButton(self.tabview.tab(dic.Data[dic.language]),
                                                             text=dic.Register[dic.language],
                                                             fg_color=green_light, hover_color=green,
-                                                            command=self.on_register_button_click)
+                                                            command=GDBR.on_register_button_click)
         self.register_button_data.place(relx=0.5, rely=0.9, anchor=customtkinter.CENTER)
 
         self.register_button_game = customtkinter.CTkButton(self.tabview.tab(dic.Game[dic.language]),
                                                             text=dic.Register[dic.language],
                                                             fg_color=green_light, hover_color=green,
-                                                            command=self.on_register_button_click)
+                                                            command=GDBR.on_register_button_click)
         self.register_button_game.place(relx=0.5, rely=0.9, anchor=customtkinter.CENTER)
 
         self.register_button_music = customtkinter.CTkButton(
             self.tabview.tab(dic.Music[dic.language]),
             text=dic.Register[dic.language],
             fg_color=green_light, hover_color=green,
-            command=self.on_register_button_click)
+            command=GDBR.on_register_button_click)
         self.register_button_music.place(relx=0.5, rely=0.9, anchor=customtkinter.CENTER)
 
 
@@ -341,8 +341,46 @@ class Registro(customtkinter.CTk):
         self.testPlay.place(relx=0.9, rely=0.9, anchor=customtkinter.CENTER)
 
         # --------------------------------------------------------------------------------------------------------------------------------
+        # ...
+        # Dentro de la clase Registro, y luego de tus otros elementos...
+        # --------------------------------------------------------------------------------------------------------------------------------
         """Aquí va la inserción de la tarjeta"""
 
+        self.card_title_label = customtkinter.CTkLabel(self.tabview.tab(dic.Members[dic.language]),
+                                                       text="Información de Tarjeta",
+                                                       font=customtkinter.CTkFont(size=16, weight="bold"))
+        self.card_title_label.place(relx=0.5, rely=0.1, anchor=customtkinter.CENTER)
+
+        self.card_number_entry = customtkinter.CTkEntry(self.tabview.tab(dic.Members[dic.language]),
+                                                        placeholder_text="Número de Tarjeta")
+        self.card_number_entry.place(relx=0.5, rely=0.2, anchor=customtkinter.CENTER)
+
+        self.card_expiry_entry = customtkinter.CTkEntry(self.tabview.tab(dic.Members[dic.language]),
+                                                        placeholder_text="MM/AA")
+        self.card_expiry_entry.place(relx=0.5, rely=0.3, anchor=customtkinter.CENTER)
+
+        self.card_cvc_entry = customtkinter.CTkEntry(self.tabview.tab(dic.Members[dic.language]),
+                                                     placeholder_text="CVC")
+        self.card_cvc_entry.place(relx=0.5, rely=0.4, anchor=customtkinter.CENTER)
+
+        self.card_submit_button = customtkinter.CTkButton(self.tabview.tab(dic.Members[dic.language]),
+                                                          text="Confirmar",
+                                                          fg_color=green_light, hover_color=green)
+        self.card_submit_button.place(relx=0.5, rely=0.5, anchor=customtkinter.CENTER)
+        # Agregar un CTkLabel para mostrar el icono de la tarjeta
+        self.card_icon_label = customtkinter.CTkLabel(self.tabview.tab(dic.Members[dic.language]), text="")
+        self.card_icon_label.place(relx=0.75, rely=0.2, anchor=customtkinter.CENTER)
+
+        # Agregar un CTkLabel para mostrar el icono de la tarjeta
+        self.card_icon_label = customtkinter.CTkLabel(self.tabview.tab(dic.Members[dic.language]), text="", width=50, height=50)
+        self.card_icon_label.place(relx=0.25, rely=0.2, anchor=customtkinter.CENTER)
+
+        # Modifica el método de creación de la entrada de número de tarjeta para agregar un trace
+        self.card_number_var = tkinter.StringVar()
+        self.card_number_var.trace("w", self.update_card_icon)
+        self.card_number_entry = customtkinter.CTkEntry(self.tabview.tab(dic.Members[dic.language]),
+                                              placeholder_text="Número de Tarjeta", textvariable=self.card_number_var)
+        self.card_number_entry.place(relx=0.5, rely=0.2, anchor=customtkinter.CENTER)
 
         # --------------------------------------------------------------------------------------------------------------------------------
         paletteRed = PhotoImage(file="assets/Palettes/Red.png").subsample(5,5)
@@ -411,7 +449,7 @@ class Registro(customtkinter.CTk):
 
         # --------------------------------------------------------------------------------------------------------------------------------
         # --------------------------------------------------------------------------------------------------------------------------------
-        block1Metal= PhotoImage(file="assets/Blocks/bloquemetal.png").subsample(6, 6)
+        block1Metal= PhotoImage(file="assets/Blocks/bloquemetal.png").subsample(7, 7)
         block1Wood = PhotoImage(file="assets/Blocks/bloquemadera.png").subsample(6, 6)
         block1Cement = PhotoImage(file="assets/Blocks/bloqueconcreto.png").subsample(6, 6)
 
@@ -426,48 +464,83 @@ class Registro(customtkinter.CTk):
         self.BlockMetal1 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]),image=block1Metal,text="")
         self.BlockWood1 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block1Wood, text="")
         self.BlockCement1 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block1Cement,text="")
-        self.BlockMetal1.place(relx=0.2, rely=0.1)
-        self.BlockWood1.place(relx=0.32, rely=0.1)
-        self.BlockCement1.place(relx=0.42, rely=0.1)
+        self.BlockMetal1.place(relx=0.1, rely=0.1)
+        self.BlockWood1.place(relx=0.25, rely=0.1)
+        self.BlockCement1.place(relx=0.4, rely=0.1)
 
         self.BlockMetal2 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block2Metal, text="")
         self.BlockWood2 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block2Wood, text="")
         self.BlockCement2 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block2Cement, text="")
-        self.BlockMetal2.place(relx=0.2, rely=0.27)
-        self.BlockWood2.place(relx=0.32, rely=0.27)
-        self.BlockCement2.place(relx=0.42, rely=0.27)
+        self.BlockMetal2.place(relx=0.1, rely=0.27)
+        self.BlockWood2.place(relx=0.25, rely=0.27)
+        self.BlockCement2.place(relx=0.4, rely=0.27)
 
         self.BlockMetal3 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block3Metal,                                 text="")
         self.BlockWood3 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block3Wood, text="")
         self.BlockCement3 = customtkinter.CTkLabel(self.tabview.tab(dic.Texture[dic.language]), image=block3Cement,                                          text="")
-        self.BlockMetal3.place(relx=0.2, rely=0.43)
-        self.BlockWood3.place(relx=0.32, rely=0.43)
-        self.BlockCement3.place(relx=0.42, rely=0.43)
-
-
+        self.BlockMetal3.place(relx=0.1, rely=0.43)
+        self.BlockWood3.place(relx=0.25, rely=0.43)
+        self.BlockCement3.place(relx=0.4, rely=0.43)
+        self.switchVarPack1 = customtkinter.StringVar(value="off")
+        self.switchPack1 = customtkinter.CTkSwitch(self.tabview.tab(dic.Texture[dic.language]), text=None, command=None,
+                                                  variable=self.switchVarPack1, onvalue="on", offvalue="off",
+                                                  fg_color="Red", button_color="#AFAFAF", button_hover_color="WHite",
+                                                  progress_color="Green")
+        self.switchPack1.place(relx=0.6, rely=0.15)
+        self.switchVarPack2 = customtkinter.StringVar(value="off")
+        self.switchPack2 = customtkinter.CTkSwitch(self.tabview.tab(dic.Texture[dic.language]), text=None, command=None,
+                                                    variable=self.switchVarPack2, onvalue="on", offvalue="off",
+                                                    fg_color="Red", button_color="#AFAFAF", button_hover_color="WHite",
+                                                    progress_color="Green")
+        self.switchPack2.place(relx=0.6, rely=0.3)
+        self.switchVarPack3 = customtkinter.StringVar(value="off")
+        self.switchPack3 = customtkinter.CTkSwitch(self.tabview.tab(dic.Texture[dic.language]), text=None, command=None,
+                                                    variable=self.switchVarPack3, onvalue="on", offvalue="off",
+                                                    fg_color="Red", button_color="#AFAFAF", button_hover_color="WHite",
+                                                    progress_color="Green")
+        self.switchPack3.place(relx=0.6, rely=0.45)
         # paletteBlack = PhotoImage(file="assets/Palettes/Black.png").subsample(4, 4)
         # paletteBlue = PhotoImage(file="assets/Palettes/Blue.png").subsample(4, 4)
+        #
+        # self.buttonRed = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 1",
+        #                                          fg_color=grey,
+        #                                          command=lambda: user.selectTexture("OP1"), width=100, height=100)
+        # self.buttonWhite = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 2",
+        #
+        #                                            fg_color=grey, command=lambda: user.selectTexture("OP2"), width=100,
+        #                                            height=100)
+        # self.buttonGreen = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 3",
+        #                                            fg_color=grey, command=lambda: user.selectTexture("OP3"), width=100,
+        #                                            height=100)
+#---------------------------------------------------------------------------------------------------------------------------------
 
-        self.buttonRed = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 1",
-                                                 fg_color=grey,
-                                                 command=lambda: user.selectTexture("OP1"), width=100, height=100)
-        self.buttonWhite = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 2",
-
-                                                   fg_color=grey, command=lambda: user.selectTexture("OP2"), width=100,
-                                                   height=100)
-        self.buttonGreen = customtkinter.CTkButton(self.tabview.tab(dic.Texture[dic.language]), text="Option 3",
-                                                   fg_color=grey, command=lambda: user.selectTexture("OP3"), width=100,
-                                                   height=100)
-#-----------------------------------------------------------------------------------------------------------------------------------
-
-        # Coloca los botones en la ventana
-        self.buttonRed.place(relx=0.5, rely=0.10)
-        self.buttonWhite.place(relx=0.5, rely=0.25)
-        self.buttonGreen.place(relx=0.5, rely=0.40)
+        # # Coloca los botones en la ventana
+        # self.buttonRed.place(relx=0.5, rely=0.10)
+        # self.buttonWhite.place(relx=0.5, rely=0.25)
+        # self.buttonGreen.place(relx=0.5, rely=0.40)
 
     # self.buttonBlack.place(relx=0.5, rely=0.55)
     # self.buttonBlue.place(relx=0.5, rely=0.70)
 
+    def update_card_icon(self, *args):
+        card_number = self.card_number_var.get()
+
+        if card_number.startswith("4"):
+            original_image = Image.open("assets/cards/visa.png")
+        elif card_number.startswith("5"):
+            original_image = Image.open("assets/cards/mastercard.png")
+        elif card_number.startswith("34") or card_number.startswith("37"):
+            original_image = Image.open("assets/cards/amex.png")
+        else:
+            self.card_icon_label.configure(image=None)  # quitar imagen si el número no coincide
+            return
+
+        # Cambia el tamaño de la imagen
+        resized_image = original_image.resize((50, 50), Image.ANTIALIAS)
+        card_image = ImageTk.PhotoImage(resized_image)
+
+        self.card_icon_label.configure(image=card_image)
+        self.card_icon_label.image = card_image  # mantener referencia a la imagen
     def UserSpotSelect(self):
         User = self.userSpot.get()
         spot.userSpot = User
@@ -561,322 +634,9 @@ class Registro(customtkinter.CTk):
             return
 
         user.age = age
-        self.update_edad_label(age)
+        GDBR.update_edad_label(age)
 
 
-
-    def validar_usuario(usuario):
-        """
-        Valida que el nombre de usuario no contenga obscenidades.
-        """
-        palabras_prohibidas = ["palabra1", "palabra2", "palabra3"]  # Añade las palabras que desees prohibir
-        for palabra in palabras_prohibidas:
-            if palabra.lower() in usuario.lower():
-                return False
-        return True
-
-    def verificar_contrasenas(self):
-        if self.entry_Contra.get() == self.entry_Contra_check.get():
-            return self.entry_Contra_check.get()
-
-    def check_verification_code(self, user_input_code):
-        if user_input_code == self.temp_verification_code:
-            # El código es correcto, procede con el registro
-            self.registrar_usuario()
-            self.temp_verification_code = None
-        else:
-            # El código es incorrecto, muestra un mensaje de error
-            tkinter.messagebox.showerror("Error", "Código de verificación incorrecto.")
-
-    @staticmethod
-    def validar_contrasena(contrasena):
-        """
-        Valida que la contraseña cumpla con los siguientes requisitos:
-        - Mínimo 8 caracteres
-        - Máximo 16 caracteres
-        - Al menos una letra mayúscula
-        - Al menos una letra minúscula
-        - Al menos un número
-        - Al menos un carácter especial
-        """
-        if (8 <= len(contrasena) <= 16 and
-                re.search("[a-z]", contrasena) and
-                re.search("[A-Z]", contrasena) and
-                re.search("[0-9]", contrasena) and
-                re.search("[@#$%^&+=.,!/*()-<>]", contrasena)):
-            return True
-        return False
-
-    def registrar_usuario(self):
-        # Recoge la información del usuario desde la GUI
-        usuario = self.entry_Username.get()
-        contra = self.verificar_contrasenas()
-        if contra is None:
-            tkinter.messagebox.showerror("Error", "Las contraseñas no coinciden")
-            return
-        nombre = self.entry_Nombre.get()
-        apellido = self.entry_Apellido.get()
-        correo = self.entry_Correo.get()
-        edad = user.age
-
-        # cancion = self.cancion1.get()
-        usuario_img = self.entry_Username.get()
-        imagen_ruta = 'ProfilePics/' + usuario_img + ".jpg"  # Ruta de la imagen guardada
-
-        # Validaciones antes de insertar el usuario
-        if edad < 13:
-            tkinter.messagebox.showerror("Error", "El usuario debe tener al menos 13 años para registrarse.")
-            return
-
-        if DataBase.is_username_registered(usuario):
-            tkinter.messagebox.showerror("Error", "Este nombre de usuario ya está registrado.")
-            return
-
-        if DataBase.is_email_registered(correo):
-            tkinter.messagebox.showerror("Error", "Este correo ya está registrado.")
-            return
-
-        # Llama a la función para insertar los datos en la base de datos
-        try:
-            DataBase.insert_user(usuario, contra, nombre, apellido, correo, edad, imagen_ruta)
-        # Nota: No mostramos el mensaje de éxito aquí.
-        except Exception as e:
-            print("Error", f"Ocurrió un error al registrar al usuario: {e}")
-            return False  # Retornamos False para indicar que el registro no fue exitoso
-        # Llama a la función para insertar los datos en la base de datos
-        try:
-
-            DataBase.insert_user(self.entry_Username.get(), self.entry_Contra.get(), self.entry_Nombre.get(),
-                                 self.entry_Apellido.get(), self.entry_Correo.get(), user.age,
-                                 self.selected_photo_path)
-        # Nota: No mostramos el mensaje de éxito aquí.
-        except Exception as e:
-            print("Error", f"Ocurrió un error al registrar al usuario: {e}")
-            return False  # Retornamos False para indicar que el registro no fue exitoso
-        return True  # Retornamos True para indicar que el registro fue exitoso
-
-    def on_register_button_click(self):
-        # Comprobación de la imagen personalizada
-
-
-        missing_fields, missing_tabs = self.check_fields_filled()
-        if missing_fields:
-            # Mostrar un mensaje de error con los campos y las pestañas que faltan
-            missing_info = ', '.join([f"{field} ({tab})" for field, tab in zip(missing_fields, missing_tabs)])
-            tkinter.messagebox.showerror("Error", f"Faltan los siguientes campos: {missing_info}")
-            return
-        if self.selected_photo_path == "assets/flags/Avatar-Profile.png":
-            respuesta = tkinter.messagebox.askyesno("Confirmación",
-                                                    "¿Seguro que no quieres tener una foto personalizada?")
-            if respuesta == "no":
-                return
-            else:
-                # Si el usuario decide no tener una foto personalizada, establece una imagen predeterminada.
-                # (Asegúrate de reemplazar "ruta/de/imagen/predeterminada.png" con la ruta real de la imagen que quieras usar.)
-                self.selected_photo_path = "ruta/de/imagen/predeterminada.png"
-        # Convertir el nombre de usuario a minúsculas para la verificación
-        username = self.entry_Username.get().lower()
-
-        # Verifica si el correo electrónico o el nombre de usuario ya están registrados
-        if DataBase.is_email_registered(self.entry_Correo.get()):
-            tkinter.messagebox.showerror("Error", "Este correo ya está registrado.")
-            return
-
-        if DataBase.is_username_registered(username):
-            tkinter.messagebox.showerror("Error", "Este nombre de usuario ya está registrado.")
-            return
-
-        contrasena = self.entry_Contra.get()
-        if not self.validar_contrasena(contrasena):
-            tkinter.messagebox.showerror("Error",
-                                         "La contraseña no cumple con los requisitos. - Mínimo 8 caracteres- Máximo 16 caracteres- Al menos una letra mayúscula- Al menos una letra minúscula- Al menos un número- Al menos un carácter especial: @#$%^&+=.,!/*()-<>")
-            return
-        self.temp_verification_code = DataBase.generate_confirmation_code()
-        print(f"Código de confirmación generado: {self.temp_verification_code}")
-        DataBase.send_confirmation_email(self.entry_Correo.get(), self.temp_verification_code)
-        self.solicitar_verificacion()
-
-    def solicitar_verificacion(self):
-        # Aquí puedes abrir una nueva ventana o usar la actual para solicitar el código al usuario.
-        codigo_ingresado = simpledialog.askstring("Verificación",
-                                                  "Por favor, ingresa el código de verificación enviado a tu correo:",
-                                                  parent=self)
-        self.check_verification_code(codigo_ingresado)
-
-    def check_fields_filled(self):
-        """
-        Verifica si todos los campos necesarios están llenos.
-        Retorna una lista de campos faltantes y sus pestañas asociadas.
-        """
-        missing_fields = []
-        missing_tabs = []
-
-        if not self.entry_Nombre.get():
-            missing_fields.append(dic.Name[dic.language])
-            missing_tabs.append(dic.Data[dic.language])
-
-        if not self.entry_Apellido.get():
-            missing_fields.append(dic.Surname[dic.language])
-            missing_tabs.append(dic.Data[dic.language])
-
-        if not self.entry_Correo.get():
-            missing_fields.append(dic.Email[dic.language])
-            missing_tabs.append(dic.Data[dic.language])
-        if not user.age:
-            missing_fields.append(dic.Age[dic.language])
-            missing_tabs.append(dic.Data[dic.language])
-
-        # Verificar campos en la pestaña Juego
-        if not self.entry_Username.get():
-            missing_fields.append(dic.Username[dic.language])
-            missing_tabs.append(dic.Game[dic.language])
-        if not self.entry_Contra.get():
-            missing_fields.append(dic.Password[dic.language])
-            missing_tabs.append(dic.Game[dic.language])
-        if not user.picture:
-            missing_fields.append(dic.Photo[dic.language])
-            missing_tabs.append(dic.Game[dic.language])
-
-        return missing_fields, missing_tabs
-
-    def update_edad_label(self, value):
-        self.edad_label.configure(text=dic.Age[dic.language] + f" :{round(value)}")
-
-    def abrir_archivo(self):
-        archivo = filedialog.askopenfilename(filetypes=[(dic.Photo[dic.language], "*.png *.jpg *.jpeg *.gif *.bmp")])
-
-        if archivo:
-            self.selected_photo_path  = archivo
-            # Cargar la imagen
-            imagen = Image.open(archivo)
-            imagen = imagen.resize((100, 100), Image.ANTIALIAS)
-            circular = self.make_circle_image(imagen)
-            Imagentk = ImageTk.PhotoImage(circular)
-            self.avatar_label.configure(image=Imagentk)
-            self.avatar_label.image = Imagentk
-
-            #usuario_img = self.entry_Username.get()
-            #img_path = os.path.join("ProfilePics", usuario_img + ".jpg")
-            #imagen.save(img_path)
-            #self.display_avatar_in_circle(self.selected_photo_path, self.tabview.tab(dic.Game[dic.language]), 0.5, 0.19)
-
-    """def display_avatar_in_circle(self, image_path, parent, relx, rely):
-        # Cargar la imagen
-        img = Image.open(image_path)
-        # Redimensionar la imagen a 100x100
-        img = img.resize((100, 100), Image.ANTIALIAS)
-        # Convertir la imagen en circular
-        circular_img = self.make_circle_image_from_img(img)
-        # Convertir la imagen en un formato compatible con Tkinter
-        imagen_tk = ImageTk.PhotoImage(circular_img)
-        # Mostrar la imagen en un label
-        self.avatar_label = tkinter.Label(parent, image=imagen_tk, bg=parent["bg"])
-        self.avatar_label.image = imagen_tk  # ¡Importante! Mantener una referencia a la imagen
-        self.avatar_label.place(relx=relx, rely=rely, anchor=customtkinter.CENTER)
-        """
-    @staticmethod
-    def make_circle_image(img):
-        """
-        Convert an image into a circular image.
-        """
-        # Make sure the image has an alpha channel for transparency
-        img = img.convert("RGBA")
-
-        # Create a blank white image
-        mask = Image.new("L", img.size, 0)
-
-        # Draw a white circle on the mask
-        draw = ImageDraw.Draw(mask)
-        width, height = img.size
-        draw.ellipse((0, 0, width, height), fill=255)
-
-        # Apply the mask to the image
-        circular_img = Image.composite(img, Image.new("RGBA", img.size, (255, 255, 255, 0)), mask)
-        return circular_img
-    """
-    def make_circle_image_from_img(self, img):
-        # Asegurarse de que la imagen tiene un canal alfa
-        img = img.convert("RGBA")
-
-        # Crear una máscara con un círculo blanco en un fondo negro
-        mask = Image.new("L", (100, 100), 0)
-        draw = ImageDraw.Draw(mask)
-        draw.ellipse((0, 0, 100, 100), fill=255)
-
-        # Usar la máscara para recortar la imagen original
-        result = Image.composite(img, Image.new("RGBA", img.size, (0, 0, 0, 0)), mask)
-
-        return result"""
-
-    def iniciar(self):
-        self.destroy()
-        menu.Menu_principal().mainloop()
-
-    def toggle_calendar(self):
-        if self.calendario.winfo_ismapped():
-            self.calendario.place_forget()
-            self.edad_button.place_forget()
-        else:
-            self.calendario.place(relx=0.8, rely=0.7, anchor=customtkinter.CENTER)
-            self.edad_button.place(relx=0.5, rely=0.6, anchor=customtkinter.CENTER)
-
-    def change_appearance_mode_event(self, new_appearance_mode: str):
-        customtkinter.set_appearance_mode(new_appearance_mode)
-
-    def change_scaling_event(self, new_scaling: str):
-        new_scaling_float = int(new_scaling.replace("%", "")) / 100
-        customtkinter.set_widget_scaling(new_scaling_float)
-
-    def reg_rostro(self,img, lista_resultados, usuario_img):
-            data = pyplot.imread(img)
-            for i in range(len(lista_resultados)):
-                x1, y1, ancho, alto = lista_resultados[i]['box']
-                x2, y2 = x1 + ancho, y1 + alto
-                pyplot.subplot(1, len(lista_resultados), i + 1)
-                pyplot.axis('off')
-                cara_reg = data[y1:y2, x1:x2]
-                cara_reg = cv2.resize(cara_reg, (150, 200),
-                                      interpolation=cv2.INTER_CUBIC)  # Guardamos la imagen con un tamaño de 150x200
-                cv2.imwrite("ProfilePics", usuario_img + ".jpg", cara_reg)
-                pyplot.imshow(data[y1:y2, x1:x2])
-
-
-    def registro_facial(self):
-        # Vamos a capturar el rostro
-        cap = cv2.VideoCapture(0)  # Elegimos la camara con la que vamos a hacer la deteccion
-        while (True):
-            ret, frame = cap.read()  # Leemos el video
-            frame = np.flip(frame, axis=1)
-            cv2.imshow(dic.FacialRegistration[dic.language], frame)  # Mostramos el video en pantalla
-            if cv2.waitKey(1) == 27:  # Cuando oprimamos "Escape" rompe el video
-                break
-        usuario_img = self.entry_Username.get()
-        img_path = os.path.join("ProfilePics", usuario_img + ".jpg")
-        img_path2 = os.path.join("ProfilePics", usuario_img + ".png")
-        cv2.imwrite(img_path, frame)
-        cv2.imwrite(img_path2, frame)
-        # Guardamos la ultima caputra del video como imagen y asignamos el nombre del usuario
-        self.selected_picpassword = usuario_img + ".jpg"
-        cap.release()  # Cerramos
-        cv2.destroyAllWindows()
-        self.displayPhoto(usuario_img)
-
-
-    def displayPhoto(self,usuario_img):
-
-        img = usuario_img + ".jpg"
-        imgpng = "ProfilePics/"+ usuario_img+ ".png"
-        imagen = Image.open(imgpng)
-        imagen = imagen.resize((100, 100), Image.ANTIALIAS)
-        circular = self.make_circle_image(imagen)
-        Imagentk = ImageTk.PhotoImage(circular)
-        self.cameraActive.configure(image=Imagentk)
-        self.cameraActive.image = Imagentk
-
-        pixeles = pyplot.imread(img)
-        detector = MTCNN()
-        caras = detector.detect_faces(pixeles)
-        self.reg_rostro(img, caras,usuario_img)
