+++ union.py
@@ -1,7 +1,11 @@
 import time
 import tkinter
 import tkinter.messagebox
+
+import customtkinter
+
 from menu import Menu_principal
+customtkinter.set_appearance_mode("dark")  # Forzar el modo oscuro
 
 
 menu = Menu_principal()
